refactor(taxonomy): reorganize taxonomy type imports and constants
